rafactor/#43: user 패키지 삭제
