+++ src/main/java/com/tookscan/tookscan/order/application/usecase/CreateAdminOrderMemoUseCase.java
@@ -1,0 +1,9 @@
+package com.tookscan.tookscan.order.application.usecase;
+
+import com.tookscan.tookscan.core.annotation.bean.UseCase;
+import com.tookscan.tookscan.order.application.dto.request.CreateAdminOrderMemoRequestDto;
+
+@UseCase
+public interface CreateAdminOrderMemoUseCase {
+    void execute(Long orderId, CreateAdminOrderMemoRequestDto requestDto);
+}
