+++ src/main/java/com/tookscan/tookscan/order/application/usecase/UpdateUserOrderScanUseCase.java
@@ -1,0 +1,9 @@
+package com.tookscan.tookscan.order.application.usecase;
+
+import com.tookscan.tookscan.core.annotation.bean.UseCase;
+import java.util.UUID;
+
+@UseCase
+public interface UpdateUserOrderScanUseCase {
+    void execute(UUID accountId, Long orderId);
+}
