--- src/main/java/com/tookscan/tookscan/order/application/user/service/CreateAdminOrderMemoService.java
@@ -1,27 +1,0 @@
-package com.tookscan.tookscan.order.application.user.service;
-
-import com.tookscan.tookscan.core.exception.error.ErrorCode;
-import com.tookscan.tookscan.core.exception.type.CommonException;
-import com.tookscan.tookscan.order.application.user.dto.request.CreateAdminOrderMemoRequestDto;
-import com.tookscan.tookscan.order.application.user.usecase.CreateAdminOrderMemoUseCase;
-import com.tookscan.tookscan.order.domain.Order;
-import com.tookscan.tookscan.order.repository.mysql.OrderRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@RequiredArgsConstructor
-public class CreateAdminOrderMemoService implements CreateAdminOrderMemoUseCase {
-
-    private final OrderRepository orderRepository;
-
-    @Override
-    @Transactional
-    public void execute(Long orderId, CreateAdminOrderMemoRequestDto requestDto) {
-        // 주문 메모 생성
-        Order order = orderRepository.findById(orderId)
-                .orElseThrow(() -> new CommonException(ErrorCode.NOT_FOUND_ORDER));
-        order.createMemo(requestDto.content());
-    }
-}
